factor out some datastore utility fns
