+++ mediachain/datastore/rpc.py
@@ -1,12 +1,10 @@
 import cbor
-import base58
 from grpc.beta import implementations
 from grpc.beta.interfaces import StatusCode
 from grpc.framework.interfaces.face.face import AbortionError
 from mediachain.proto import Datastore_pb2  # pylint: disable=no-name-in-module
-from mediachain.proto import Types_pb2  # pylint: disable=no-name-in-module
-from mediachain.datastore.data_objects import MultihashReference
-
+from mediachain.datastore.utils import rpc_ref, multihash_ref, \
+    bytes_for_object, object_for_bytes
 
 TIMEOUT_SECS = 120
 
@@ -17,10 +15,10 @@ class RpcDatastore(object):
         self.rpc = Datastore_pb2.beta_create_DatastoreService_stub(channel)
 
     def put(self, data_object, timeout=TIMEOUT_SECS):
-        byte_string = object_bytes(data_object)
-        req = Datastore_pb2.DataObject(bytes=byte_string)
+        byte_string = bytes_for_object(data_object)
+        req = Datastore_pb2.DataObject(data=byte_string)
         ref = self.rpc.put(req, timeout)
-        return MultihashReference.from_base58(ref.reference)
+        return multihash_ref(ref)
 
     def get(self, ref, timeout=TIMEOUT_SECS):
         ref = rpc_ref(ref)
@@ -30,38 +28,4 @@ class RpcDatastore(object):
             if e.code == StatusCode.NOT_FOUND:
                 return None
             raise
-        try:
-            return cbor.loads(obj_bytes)
-        except ValueError:
-            return obj_bytes
-
-
-def rpc_ref(ref):
-    ref_str = ref
-    try:
-        ref_str = ref.multihash_base58()
-    except AttributeError:
-        pass
-    try:
-        ref_str = ref.reference
-    except AttributeError:
-        pass
-    try:
-        ref_str = base58.b58encode(ref['@link'])
-    except (KeyError, ValueError, TypeError):
-        pass
-    return Types_pb2.MultihashReference(reference=ref_str)
-
-
-def object_bytes(obj):
-    try:
-        return obj.to_cbor_bytes()
-    except AttributeError:
-        pass
-
-    if isinstance(object, dict):
-        try:
-            return cbor.dumps(obj)
-        except ValueError:
-            pass
-    return obj
+        return object_for_bytes(obj_bytes)
